# Conflicts:
#	data/assets/scene/solarsystem/missions/apollo/11/apollo11.asset
#	data/assets/scene/solarsystem/planets/neptune/irregular_retrograde_moons.asset
#	data/assets/scene/solarsystem/planets/saturn/dione/dione.asset
#	data/assets/scene/solarsystem/planets/saturn/dione/trail.asset
#	data/assets/scene/solarsystem/planets/saturn/enceladus/enceladus.asset
#	data/assets/scene/solarsystem/planets/saturn/enceladus/trail.asset
#	data/assets/scene/solarsystem/planets/saturn/hyperion/hyperion.asset
#	data/assets/scene/solarsystem/planets/saturn/hyperion/trail.asset
#	data/assets/scene/solarsystem/planets/saturn/iapetus/iapetus.asset
#	data/assets/scene/solarsystem/planets/saturn/iapetus/trail.asset
#	data/assets/scene/solarsystem/planets/saturn/mimas/trail.asset
#	data/assets/scene/solarsystem/planets/saturn/minor/gallic_group.asset
#	data/assets/scene/solarsystem/planets/saturn/minor/inuit_group.asset
#	data/assets/scene/solarsystem/planets/saturn/minor/norse_group.asset
#	data/assets/scene/solarsystem/planets/saturn/rhea/trail.asset
#	data/assets/scene/solarsystem/planets/saturn/tethys/trail.asset
#	data/assets/scene/solarsystem/planets/saturn/titan/titan.asset
#	data/assets/scene/solarsystem/planets/saturn/titan/trail.asset
#	data/assets/scene/solarsystem/planets/uranus/irregular_retrograde_moons.asset
#	data/assets/util/procedural_globe.asset
#	include/openspace/navigation/orbitalnavigator.h
#	modules/globebrowsing/CMakeLists.txt
#	modules/globebrowsing/shaders/geojson_fs.glsl
#	modules/globebrowsing/shaders/geojson_vs.glsl
#	modules/touch/src/touchinteraction.cpp
#	openspace.cfg
#	src/navigation/orbitalnavigator.cpp

+++ modules/skybrowser/src/wwtdatahandler.cpp
@@ -171,7 +171,7 @@ namespace {
         return true;
     }
 
-    std::optional<openspace::ImageData> 
+    std::optional<openspace::ImageData>
     loadImageFromNode(const tinyxml2::XMLElement* node, const std::string& collection)
     {
         using namespace openspace;
@@ -330,7 +330,7 @@ void WwtDataHandler::loadImages(const std::string& root,
         }
     );
     // Set the identifiers to the correct order
-    for (int i = 0; i < _imageVector.size(); i++) {
+    for (size_t i = 0; i < _imageVector.size(); i++) {
         _images[_imageVector[i].imageUrl].identifier = std::to_string(i);
     }
